Create a component for outdoor/weather function
